Feature/filter advertisement

+++ selling-used-items-app-backend/selling-used-items-app-backend/Controllers/AdvertisementController.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using selling_used_items_app_backend.Model;
 using selling_used_items_app_backend.Service;
+using selling_used_items_app_backend.Validator.AdvertisementValidator;
+using System.ComponentModel.DataAnnotations;
 
 namespace selling_used_items_app_backend.Controllers
 {
@@ -8,11 +10,17 @@ namespace selling_used_items_app_backend.Controllers
     [Route("api/advertisements")]
     public class AdvertisementController : ControllerBase
     {
-        private readonly AdvertisementService _advertisementService;
+        private readonly IAdvertisementService _advertisementService;
+        private readonly AdvertisementCreateValidator _advertisementCreateValidator;
+        private readonly AdvertisementUpdateValidator _advertisementUpdateValidator;
+        private readonly AdvertisementDeleteValidator _advertisementDeleteValidator;
 
-        public AdvertisementController(AdvertisementService advertisementService)
+        public AdvertisementController(IAdvertisementService advertisementService, AdvertisementCreateValidator advertisementCreateValidator, AdvertisementUpdateValidator advertisementUpdateValidator, AdvertisementDeleteValidator advertisementDeleteValidator)
         {
-            _advertisementService = advertisementService ?? throw new ArgumentNullException(nameof(advertisementService));
+            _advertisementService = advertisementService;
+            _advertisementCreateValidator = advertisementCreateValidator;
+            _advertisementUpdateValidator = advertisementUpdateValidator;
+            _advertisementDeleteValidator = advertisementDeleteValidator;
         }
 
         [HttpGet]
@@ -23,9 +31,9 @@ namespace selling_used_items_app_backend.Controllers
         }
 
         [HttpGet("{id}")]
-        public ActionResult<Advertisement> GetById(int id)
+        public ActionResult<Advertisement> Get(int id)
         {
-            var advertisement = _advertisementService.GetById(id);
+            var advertisement = _advertisementService.Get(id);
             if (advertisement == null)
             {
                 return NotFound();
@@ -36,8 +44,13 @@ namespace selling_used_items_app_backend.Controllers
         [HttpPost]
         public IActionResult Create(Advertisement advertisement)
         {
+            var validationResult = _advertisementCreateValidator.ValidateAdvertisement(advertisement);
+            if (validationResult != ValidationResult.Success)
+            {
+                return BadRequest(validationResult.ErrorMessage);
+            }
             _advertisementService.Create(advertisement);
-            return CreatedAtAction(nameof(GetById), new { id = advertisement.id }, advertisement);
+            return CreatedAtAction(nameof(Get), new { id = advertisement.id }, advertisement);
         }
 
         [HttpPut("{id}")]
@@ -45,9 +58,13 @@ namespace selling_used_items_app_backend.Controllers
         {
             if (id != advertisement.id)
             {
-                return BadRequest();
+                return BadRequest("ID in the request path does not match the ID in the request body.");
             }
-
+            var validationResult = _advertisementCreateValidator.ValidateAdvertisement(advertisement);
+            if (validationResult != ValidationResult.Success)
+            {
+                return BadRequest(validationResult.ErrorMessage);
+            }
             _advertisementService.Update(advertisement);
             return NoContent();
         }
@@ -55,8 +72,20 @@ namespace selling_used_items_app_backend.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
+            var validationResult = _advertisementDeleteValidator.ValidateAdvertisement(id);
+            if (validationResult != ValidationResult.Success)
+            {
+                return NotFound(validationResult.ErrorMessage);
+            }
             _advertisementService.Delete(id);
             return NoContent();
+        }
+
+        [HttpGet("search")]
+        public ActionResult<IEnumerable<Advertisement>> Search(string name = null, char? firstLetter = null, decimal? startPrice = null, decimal? endPrice = null)
+        {
+            var advertisements = _advertisementService.Search(name, firstLetter, startPrice, endPrice);
+            return Ok(advertisements);
         }
     }
 }
